Refactor runtime implementation details into separate files
Since runtime is source lang and implementation specific but code generation pass is intended to be generic
