+++ src/passes/llvm-ir/code_generation.cc
@@ -1,4 +1,5 @@
 #include "../../llvm-lang.hh"
+#include "../../runtime.hh"
 #include "../llvm_utils.hh"
 #include "trieste/pass.h"
 #include "trieste/rewrite.h"
@@ -17,11 +18,6 @@ namespace llvmir {
 
   using namespace trieste;
   using namespace llvm;
-
-  // Helper function prototypes
-  void genExternalFunctions(std::shared_ptr<LLVMIRContext> context);
-  void genPrintInt(std::shared_ptr<LLVMIRContext> context);
-  void genPrintBool(std::shared_ptr<LLVMIRContext> context);
 
   /**
    * @brief
@@ -637,26 +633,9 @@ namespace llvmir {
           }};
 
     pass.pre([ctx](Node) {
-      // TODO: Refactor these to separate file to decouple code gen from source
-      // language.
-      /**
-       * External functions
-       */
+
       genExternalFunctions(ctx);
-
-      /**
-       * Internal native functions
-       */
-      genPrintInt(ctx);
-      genPrintBool(ctx);
-
-      // Declare a closure type
-      StructType* ClosureTy =
-        StructType::create(ctx->llvm_context, "ClosureTy");
-      ClosureTy->setBody(
-        {ctx->builder.getPtrTy(), ctx->builder.getPtrTy()}, false);
-
-      ctx->types["ClosureTy"] = ClosureTy;
+      genRuntimeFunctions(ctx);
 
       return 0;
     });
@@ -696,103 +675,6 @@ namespace llvmir {
     });
 
     return pass;
-  }
-
-  /**
-   * Generates imports of external functions needed for LLVM IR generation.
-   * @param context LLVM IR context
-   */
-  void genExternalFunctions(std::shared_ptr<LLVMIRContext> context) {
-    // printf(char*) -> i32
-    FunctionType* printfFunctionType = FunctionType::get(
-      context->builder.getInt32Ty(),
-      {context->builder.getInt8Ty()->getPointerTo()},
-      true);
-    context->llvm_module.getOrInsertFunction("printf", printfFunctionType);
-
-    // malloc(i64) -> ptr
-    FunctionType* mallocFunctionType = FunctionType::get(
-      context->builder.getPtrTy(), {context->builder.getInt64Ty()}, false);
-    context->llvm_module.getOrInsertFunction("malloc", mallocFunctionType);
-  }
-
-  /**
-   * Generates a function that prints an integer value.
-   * @param context LLVM IR context
-   */
-  void genPrintInt(std::shared_ptr<LLVMIRContext> context) {
-    // Formatstring needed to make calls to printf
-    auto formatStrInt = context->builder.CreateGlobalStringPtr(
-      "%d\n", "formatStrInt", 0, &context->llvm_module);
-
-    // printInt(i32) -> i32
-    FunctionType* theFunctionType = FunctionType::get(
-      context->builder.getInt32Ty(), {context->builder.getInt32Ty()}, false);
-
-    // FIXME: native functions use an internal name which must not be shadowed
-    // by the user. Need to use a name that is not allowable in the language.
-    std::string functionName = "native$printInt";
-    Function* theFunction = Function::Create(
-      theFunctionType,
-      Function::LinkageTypes::ExternalLinkage,
-      functionName,
-      context->llvm_module);
-    theFunction->setCallingConv(CallingConv::C);
-    context->registers[functionName] = theFunction;
-
-    BasicBlock* functionEntryBlock =
-      BasicBlock::Create(context->llvm_context, "entry", theFunction);
-    context->builder.SetInsertPoint(functionEntryBlock);
-
-    Argument* arg = theFunction->arg_begin();
-    arg->setName("intToPrint");
-
-    Function* printfFunc = context->llvm_module.getFunction("printf");
-    context->builder.CreateCall(printfFunc, {formatStrInt, arg});
-
-    context->builder.CreateRet(arg);
-
-    verifyFunction(*theFunction, &llvm::errs());
-  }
-  /**
-   * Generates a function that prints a boolean value.
-   * @param context LLVM IR context
-   */
-  void genPrintBool(std::shared_ptr<LLVMIRContext> context) {
-    // String needed to make calls to printf
-    auto strBoolTrue = context->builder.CreateGlobalStringPtr(
-      "true\n", "strBoolTrue", 0, &context->llvm_module);
-    auto strBoolFalse = context->builder.CreateGlobalStringPtr(
-      "false\n", "strBoolFalse", 0, &context->llvm_module);
-
-    // printInt(i1) -> i1
-    FunctionType* theFunctionType = FunctionType::get(
-      context->builder.getInt1Ty(), {context->builder.getInt1Ty()}, false);
-
-    std::string functionName = "native$printBool";
-    Function* theFunction = Function::Create(
-      theFunctionType,
-      Function::LinkageTypes::ExternalLinkage,
-      functionName,
-      context->llvm_module);
-    theFunction->setCallingConv(CallingConv::C);
-    context->registers[functionName] = theFunction;
-
-    BasicBlock* functionEntryBlock =
-      BasicBlock::Create(context->llvm_context, "entry", theFunction);
-    context->builder.SetInsertPoint(functionEntryBlock);
-
-    Argument* arg = theFunction->arg_begin();
-    arg->setName("boolToPrint");
-
-    Function* printfFunc = context->llvm_module.getFunction("printf");
-    context->builder.CreateCall(
-      printfFunc,
-      {context->builder.CreateSelect(arg, strBoolTrue, strBoolFalse)});
-
-    context->builder.CreateRet(arg);
-
-    verifyFunction(*theFunction, &llvm::errs());
   }
 
 }
